MMPCS/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
boost=1.74.0=py310h7c3ba0c_5
boost-cpp=1.74.0=h6cacc03_7
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.7.22=hbcca054_0
cairo=1.16.0=ha00ac49_1009
certifi=2023.7.22=pyhd8ed1ab_0
charset-normalizer=3.2.0=pypi_0
cmake=3.27.4.1=pypi_0
cycler=0.11.0=pyhd8ed1ab_0
filelock=3.12.4=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.1=hef1e5e3_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.10.4=hca18f0e_2
fsspec=2023.9.2=pypi_0
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h0b41bf4_3
greenlet=2.0.2=py310hc6cd4ac_1
huggingface-hub=0.17.3=pypi_0
icu=69.1=h9c3ff4c_0
idna=3.4=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h0b41bf4_3
kiwisolver=1.4.5=py310hd41b1e2_0
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=18_linux64_openblas
libcblas=3.9.0=18_linux64_openblas
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_0
libgfortran-ng=13.2.0=h69a702a_0
libgfortran5=13.2.0=ha4646dd_0
libglib=2.78.0=hebfc3b9_0
libgomp=13.2.0=h807b86a_0
libiconv=1.17=h166bdaf_0
liblapack=3.9.0=18_linux64_openblas
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.24=pthreads_h413a1c8_0
libpng=1.6.39=h753d276_0
libsqlite=3.43.0=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_0
libtiff=4.5.0=h6adf6a1_2
libuuid=2.38.1=h0b41bf4_0
libwebp=1.2.4=h1daa5a0_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h0b41bf4_0
libxml2=2.9.14=haae042b_4
libzlib=1.2.13=hd590300_5
lit=16.0.6=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib-base=3.4.3=py310h23f4a51_2
mpmath=1.3.0=pypi_0
ncurses=6.4=hcb278e6_0
networkx=3.1=pypi_0
numpy=1.25.2=py310ha4c1d20_0
openssl=3.1.2=hd590300_0
packaging=23.1=pypi_0
pandas=2.1.0=py310hcc13569_0
pcre2=10.40=hc3806b6_0
pillow=9.4.0=py310h6a678d5_0
pip=23.2.1=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
psutil=5.9.5=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pycairo=1.24.0=py310hda9f760_0
pyg-lib=0.2.0+pt20cu118=pypi_0
pyparsing=3.1.1=pyhd8ed1ab_0
python=3.10.12=hd12c33a_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.10=3_cp310
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=6.0.1=pypi_0
rdkit=2022.03.5=py310h1c297d8_0
readline=8.2=h8228510_1
regex=2023.8.8=pypi_0
reportlab=3.6.12=py310h5eee18b_0
requests=2.31.0=pypi_0
safetensors=0.3.3=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.2=pypi_0
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sqlalchemy=2.0.20=py310h2372a71_0
sympy=1.12=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h27826a3_0
tokenizers=0.13.3=pypi_0
torch=2.0.0+cu118=pypi_0
torch-cluster=1.6.1+pt20cu118=pypi_0
torch-geometric=2.3.1=pypi_0
torch-scatter=2.1.1+pt20cu118=pypi_0
torch-sparse=0.6.17+pt20cu118=pypi_0
torch-spline-conv=1.2.2+pt20cu118=pypi_0
torchaudio=2.0.1+cu118=pypi_0
torchvision=0.15.1+cu118=pypi_0
tornado=6.3.3=py310h2372a71_0
tqdm=4.66.1=pypi_0
transformers=4.33.2=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=2.0.4=pypi_0
wheel=0.41.2=pyhd8ed1ab_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.6=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.11=hd590300_0
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h166bdaf_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
